RQ-216 Bypass CORS issue by triggering request from extension
