+++ app/src/views/features/apis/apis.scss
@@ -36,7 +36,7 @@
       min-height: 0;
 
       .gutter:hover {
-        background-color: #818cf8;
+        background-color: #1e69ff;
       }
     }
